(refactor): rename localStorageAdaptor to asyncLocalStorage
- because that makes a bit more sense / is more intuitive
